+++ Magic/CHANGELOG.md
@@ -19,7 +19,8 @@
  - Hex color support in all messages in-game using an `#AA22FF` format
 
 ## General Changes and Additions
-
+ 
+ - The `spawn` mob trigger now only fires on initial spawn. Persistent mobs can look for `respawn`.
  - Add `prevent_teleport` mob option, and a `contained_enderman` mob for arena battles
  - Add `use_parent_target_location` and `use_parent_source_location` ChangeContext options
  - Add `teleport_target` Recall option